requirements_ss4rec.txt
# SS4Rec Dependencies
# Core ML framework dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# State Space Model dependencies
mamba-ssm==2.2.2
s5-pytorch==0.2.1
causal-conv1d>=1.2.0

# RecBole framework (optional - for official implementation compatibility)
recbole==1.2.0

# Experiment tracking and monitoring
wandb>=0.15.0
tensorboard>=2.10.0

# Data processing
tqdm>=4.64.0
pyyaml>=6.0

# Evaluation and metrics
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.900
ruff>=0.0.200

# Existing project dependencies (maintain compatibility)
streamlit>=1.25.0
plotly>=5.10.0
fastapi>=0.95.0
uvicorn>=0.20.0